# static/swagger.json
{
  "swagger": "2.0.0",
  "info": {
    "description": "Python project to export from Bigquery to GCS",
    "version": "1.0.0",
    "title": "Cloud Run Python",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "tags": [
    {
      "name": "Export to GCS",
      "description": "API for exporting Bigquery table to GCS"
    }
  ],
  "paths": {
    "/export/{dataset_id}/{table_id}": {
      "parameters": [
        {
          "name": "dataset_id",
          "in": "path",
          "required": true,
          "description": "ID the dataset in bigquery",
          "type": "string"
        },
        {
          "name": "table_id",
          "in": "path",
          "required": true,
          "description": "ID the table under the dataset in bigquery",
          "type": "string"
        }
      ],
      "post": {
        "tags": [
          "BQ to GCS"
        ],
        "summary": "Export Bigquery table to a single CSV file",
        "requestBody": {
          "description": "Export file post request",
          "required": true,
          "content": {
            "application/json": {
                "$ref": "#/components/schemas/Input"
            }
          }
        },
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
              "$ref": "#/components/schemas/Output"
          },
          "400": {
            "description": "Failed. Bad post data."
          },
          "404": {
            "description": "Failed. Wrong Arguments."
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Output": {
        "type": "object",
        "properties": {
          "path": {
            "type": "string",
            "format": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "Input": {
        "type": "object",
        "properties": {
          "project": {
            "type": "string",
            "format": "string"
          },
          "bucket": {
            "type": "string",
            "format": "string"
          },
          "location": {
            "type": "string",
            "format": "string"
          },
          "with_header": {
            "type": "string",
            "format": "string"
          }
        }
      }
    }
  }
}